moved timestamp parsing to consts
